Remove support for using SoX instead of FFmpeg

+++ LoopingAudioConverter/Options.cs
@@ -55,8 +55,8 @@ namespace LoopingAudioConverter {
 		public int? SampleRate { get; set; }
 		public decimal? AmplifydB { get; set; }
 		public decimal? AmplifyRatio { get; set; }
-		public decimal? PitchSemitones { get; set; }
-		public decimal? TempoRatio { get; set; }
+		public double? PitchSemitones { get; set; }
+		public double? TempoRatio { get; set; }
 		public ChannelSplit ChannelSplit { get; set; }
 		public ExporterType ExporterType { get; set; }
 		public string MP3EncodingParameters { get; set; }
